refactor(corebundle): split shared_deps.go into per-concern files (D2 part 1)
Pure file move, zero callsite churn. Apply PR-A66 actual范式 (flat struct +
per-concern file split) to cmd/corebundle/shared_deps.go (587 → 232 lines):

- shared_deps_validate.go (203 lines): Validate + 4 validate* + isLoopbackBindAddr
  startup invariant suite extracted to its own file. New site for adding
  invariant rules without scrolling past struct/build helpers.

- shared_deps_build.go (135 lines): buildSharedMetricsDeps / buildSharedReplayDeps
  / resolveListenerAddrs / closeRedisClientAfterFailedLoad / adapterInfoForSharedDeps
  + intermediate sharedReplayDeps / sharedMetricsDeps types extracted as the
  "construct-time helpers" group.

- shared_deps.go (232 lines): retains SharedDeps struct (flat — see updated
  godoc citing PR-A66 rationale), SampleVerbosePlaceholder const, and
  LoadSharedDepsFromEnv composition root.

Closes D2 (PR-V1-SHARED-DEPS-SPLIT) part 1 of 2; bundle.go split follows
in next commit per shipping plan.

ref: runtime/bootstrap/bootstrap.go:50-71 (flat struct + per-concern file
split rationale, fx/kratos citations)
