+++ cmd/corebundle/shared_deps.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"log/slog"
-	"net"
 	"net/http"
 	"os"
 	"strings"
@@ -16,8 +13,6 @@ import (
 	adapterredis "github.com/ghbvf/gocell/adapters/redis"
 	"github.com/ghbvf/gocell/kernel/clock"
 	"github.com/ghbvf/gocell/kernel/idempotency"
-	"github.com/ghbvf/gocell/pkg/errcode"
-	"github.com/ghbvf/gocell/runtime/auth"
 	"github.com/ghbvf/gocell/runtime/bootstrap"
 	"github.com/ghbvf/gocell/runtime/eventbus"
 	obmetrics "github.com/ghbvf/gocell/runtime/observability/metrics"
@@ -30,9 +25,20 @@ import (
 // SharedDeps is passed directly to BuildApp and each CellModule.Provide,
 // giving type-safe access to all cross-cutting fields without type-assertion.
 //
+// Fields are flat (no concern-grouped sub-structs): SharedDeps is a
+// composition-root bag whose fields cross consumer boundaries (Clock /
+// Topology / SharedPGPool consumed by every Cell module). Forcing a sub-struct
+// layout would make those cross-cutting consumptions look like boundary
+// violations when in fact they are the natural shape of a composition root.
+// Per-concern *file* split is in shared_deps_build.go (build helpers) and
+// shared_deps_validate.go (startup invariants); the struct itself stays flat,
+// matching runtime/bootstrap/bootstrap.go which adopted the same trade-off.
+//
 // ref: uber-go/fx fx.Supply — shared values provided once to all modules.
 // ref: kubernetes/kubernetes cmd/kube-apiserver/app/options/validation.go —
 // all required fields validated in one place before startup.
+// ref: runtime/bootstrap/bootstrap.go:50-71 — flat struct + per-concern file
+// split rationale (PR-A66 BOOTSTRAP-STRUCT-DECOMPOSE).
 type SharedDeps struct {
 	// Clock is the single root clock instance threaded through every adapter,
 	// service, and middleware constructed by BuildApp / its module builders.
@@ -162,126 +168,6 @@ type SharedDeps struct {
 	keyProviderMetricCollectors []prom.Collector
 }
 
-type sharedReplayDeps struct {
-	RedisClient         *adapterredis.Client
-	NonceStore          auth.NonceStore
-	ConsumerClaimer     idempotency.Claimer
-	ConsumerClaimerKind consumerClaimerKind
-}
-
-type sharedMetricsDeps struct {
-	PromStack            promStack
-	ConfigEventCollector obmetrics.ConfigEventCollector
-}
-
-func buildSharedMetricsDeps() (sharedMetricsDeps, error) {
-	ps, err := buildPromStack()
-	if err != nil {
-		return sharedMetricsDeps{}, err
-	}
-	configEventCollector, err := obmetrics.NewProviderConfigEventCollector(ps.metricProvider)
-	if err != nil {
-		return sharedMetricsDeps{}, fmt.Errorf("build config event metrics collector: %w", err)
-	}
-	return sharedMetricsDeps{
-		PromStack:            ps,
-		ConfigEventCollector: configEventCollector,
-	}, nil
-}
-
-func buildSharedReplayDeps(ctx context.Context, topo bootstrap.Topology, clk clock.Clock) (sharedReplayDeps, error) {
-	redisResult, err := buildRedisClient(ctx, topo)
-	if err != nil {
-		return sharedReplayDeps{}, err
-	}
-	redisClient := redisResult.Client
-	loaded := false
-	defer func() {
-		if !loaded {
-			closeRedisClientAfterFailedLoad(ctx, redisClient)
-		}
-	}()
-
-	nonceStore, err := buildServiceNonceStore(topo, redisClient, clk)
-	if err != nil {
-		return sharedReplayDeps{}, err
-	}
-	claimer, claimerKind, err := buildConsumerClaimer(topo, redisClient, clk)
-	if err != nil {
-		return sharedReplayDeps{}, err
-	}
-
-	loaded = true
-	return sharedReplayDeps{
-		RedisClient:         redisClient,
-		NonceStore:          nonceStore,
-		ConsumerClaimer:     claimer,
-		ConsumerClaimerKind: claimerKind,
-	}, nil
-}
-
-// resolveListenerAddrs returns primary / internal / health bind addresses,
-// applying default ports when the matching env var is unset:
-//
-//   - primary  → `:8080`
-//   - internal → `127.0.0.1:9090` (loopback by default; service-token gated
-//     in every mode; operators binding to a VPC interface must set
-//     GOCELL_HTTP_INTERNAL_ADDR explicitly)
-//   - health   → `127.0.0.1:9091` (separate loopback port; real-mode
-//     PodIP/Service probes must set a Pod-reachable bind such as `:9091`,
-//     or explicitly opt into same-netns access with GOCELL_HTTP_HEALTH_LOCAL_ONLY=1)
-func resolveListenerAddrs() (primary, internal, health string) {
-	primary = os.Getenv("GOCELL_HTTP_PRIMARY_ADDR")
-	if primary == "" {
-		primary = ":8080"
-	}
-	internal = os.Getenv("GOCELL_HTTP_INTERNAL_ADDR")
-	if internal == "" {
-		internal = "127.0.0.1:9090"
-	}
-	health = os.Getenv("GOCELL_HTTP_HEALTH_ADDR")
-	if health == "" {
-		health = "127.0.0.1:9091"
-	}
-	return
-}
-
-// closeRedisClientAfterFailedLoad is the single source of truth for "close
-// Redis with nil-safe + slog warn". Two callers, both following the same
-// `if !loaded { close }` defer pattern (one inside buildSharedReplayDeps for
-// inner-construction failure, one in buildSharedDeps for outer-composition
-// failure after replay deps are already attached). The structure is mirrored
-// at both sites so the two scopes can be visually compared in one read.
-func closeRedisClientAfterFailedLoad(ctx context.Context, client *adapterredis.Client) {
-	if client == nil {
-		return
-	}
-	if closeErr := client.Close(ctx); closeErr != nil {
-		slog.Warn("corebundle: failed to close Redis client after startup validation failure",
-			slog.String("error", closeErr.Error()))
-	}
-}
-
-func adapterInfoForSharedDeps(shared *SharedDeps) map[string]string {
-	info := shared.Topology.AdapterInfo()
-	redisState := "not-configured"
-	if shared.RedisClient != nil {
-		redisState = "configured"
-	}
-	nonceStoreKind := string(auth.NonceStoreKindNoop)
-	if shared.InternalGuard != nil && shared.InternalGuard.NonceStore() != nil {
-		nonceStoreKind = string(shared.InternalGuard.NonceStore().Kind())
-	}
-	claimerKind := string(shared.ConsumerClaimerKind)
-	if claimerKind == "" {
-		claimerKind = string(consumerClaimerKindUnknown)
-	}
-	info["redis"] = redisState
-	info["service_token_nonce_store"] = nonceStoreKind
-	info["outbox_consumer_claimer"] = claimerKind
-	return info
-}
-
 // SampleVerbosePlaceholder is the literal placeholder shipped in .env.example so
 // `cp .env.example .env && go run ./cmd/corebundle` works without first
 // minting a secret. validateControlPlane rejects this exact value in
@@ -289,200 +175,7 @@ func adapterInfoForSharedDeps(shared *SharedDeps) map[string]string {
 // high-entropy token. Exposed (capitalised) so example/test code and the
 // regression test in shared_deps_test.go reference one source of truth.
 // production deployments must mint their own high-entropy token.
-//
-
 const SampleVerbosePlaceholder = "dev-readyz-verbose-token-change-me"
-
-// Validate is the startup invariant check for all cross-cutting dependencies.
-// Storage-specific invariants (PGResource, cursor codecs, HMAC key) are checked
-// inside the corresponding CellModule.Provide, not here.
-//
-// ref: kubernetes/kubernetes cmd/kube-apiserver/app/options/validation.go —
-// validates all fields before any component is constructed.
-func (d *SharedDeps) Validate() error {
-	if d == nil {
-		return errcode.New(errcode.KindInvalid, errcode.ErrValidationFailed, "SharedDeps: nil receiver")
-	}
-	errs := d.validateCore()
-	errs = append(errs, d.validateVerboseEndpoint()...)
-	errs = append(errs, d.validateHealthReachability()...)
-	errs = append(errs, d.validateInternalListenerGuard()...)
-	errs = append(errs, d.validateControlPlane()...)
-	return errors.Join(errs...)
-}
-
-// validateVerboseEndpoint enforces that every adapter mode either configures
-// a verbose token or explicitly waives the endpoint. The previous dev-mode
-// fallback (unset env var => verbose open) was removed in PR-A35 so a
-// forgotten GOCELL_READYZ_VERBOSE_TOKEN in dev cannot silently expose cell
-// topology to anyone who can reach the port.
-func (d *SharedDeps) validateVerboseEndpoint() []error {
-	if d.VerboseDisabled {
-		// Both set is not a hard validation failure — VerboseDisabled
-		// wins, Handler will serve the plain aggregate body regardless of
-		// the token. But it is almost certainly a misconfiguration: the
-		// operator either wanted token-gated access (drop the DISABLED
-		// flag) or wanted to waive verbose entirely (unset the TOKEN).
-		// Surface it as a Warn so operators can spot it in startup logs.
-		if d.VerboseToken != "" {
-			slog.Warn("GOCELL_READYZ_VERBOSE_TOKEN is set but GOCELL_READYZ_VERBOSE_DISABLED=1 overrides it; " +
-				"the token will not be enforced. Drop one of the two env vars to remove the ambiguity.")
-		}
-		return nil
-	}
-	if d.VerboseToken != "" {
-		return nil
-	}
-	return []error{errcode.New(errcode.KindInternal, errcode.ErrControlplaneVerboseTokenMissing,
-		"GOCELL_READYZ_VERBOSE_TOKEN must be set (or "+
-			"GOCELL_READYZ_VERBOSE_DISABLED=1 to waive the verbose endpoint) "+
-			"so /readyz?verbose is never anonymous")}
-}
-
-// validateCore collects missing-field errors for dependencies required in
-// every topology.
-func (d *SharedDeps) validateCore() []error {
-	var errs []error
-	missing := func(field string) {
-		errs = append(errs, errcode.New(errcode.KindInvalid, errcode.ErrValidationFailed,
-			"SharedDeps field must be set",
-			errcode.WithInternal(fmt.Sprintf("field=%s", field))))
-	}
-	// Clock is the single root clock instance threaded through every adapter,
-	// service, and middleware. Required: a nil here would surface as a deeper
-	// runtime panic (kernel/clock.MustHaveClock or constructors that re-validate).
-	// Validate at startup so misconfiguration fails before any subsystem starts.
-	if d.Clock == nil {
-		missing("Clock")
-	}
-	if d.JWTDeps.issuer == nil {
-		missing("JWTDeps.issuer")
-	}
-	if d.JWTDeps.verifier == nil {
-		missing("JWTDeps.verifier")
-	}
-	if d.PromStack.registry == nil {
-		missing("PromStack.registry")
-	}
-	if d.PromStack.hookObserver == nil {
-		missing("PromStack.hookObserver")
-	}
-	if d.PromStack.metricProvider == nil {
-		missing("PromStack.metricProvider")
-	}
-	if d.EventBus == nil {
-		missing("EventBus")
-	}
-	if d.ConfigEventCollector == nil {
-		missing("ConfigEventCollector")
-	}
-	if d.ConsumerClaimer == nil {
-		missing("ConsumerClaimer")
-	}
-	return errs
-}
-
-// validateControlPlane collects errors for the production control-plane gate
-// (tokens + guard required whenever real keys are in use).
-func (d *SharedDeps) validateControlPlane() []error {
-	if !d.Topology.RequireProductionControlPlane() {
-		return nil
-	}
-	var errs []error
-	// The unconditional /readyz?verbose invariant is now enforced by
-	// validateVerboseEndpoint in every mode. Production additionally forbids
-	// waiving the endpoint: a "real" deployment that still sets
-	// GOCELL_READYZ_VERBOSE_DISABLED=1 is almost certainly a misconfiguration
-	// and would leave operators without a token-gated diagnostic path.
-	if d.VerboseDisabled {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneVerboseTokenMissing,
-			"GOCELL_READYZ_VERBOSE_DISABLED=1 is not allowed in adapter mode "+
-				"\"real\"; production must keep the token-gated verbose endpoint "+
-				"available for on-call diagnostics"))
-	}
-	if d.VerboseToken == SampleVerbosePlaceholder {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneVerboseTokenSample,
-			"GOCELL_READYZ_VERBOSE_TOKEN is set to the .env.example placeholder; a production deploy must mint its own high-entropy secret",
-			errcode.WithInternal(fmt.Sprintf("placeholder=%q", SampleVerbosePlaceholder))))
-	}
-	if d.MetricsToken == "" {
-		errs = append(errs, errcode.New(errcode.KindInvalid, errcode.ErrValidationFailed,
-			"GOCELL_METRICS_TOKEN must be set in adapter mode \"real\" to "+
-				"prevent anonymous /metrics exposure; scrapers must send "+
-				"X-Metrics-Token header"))
-	}
-	if d.InternalGuard == nil {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneServiceSecretMissing,
-			"GOCELL_SERVICE_SECRET must be set in adapter mode \"real\" to protect /internal/v1/*"))
-	} else if ns := d.InternalGuard.NonceStore(); ns == nil {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneNonceStoreMissing,
-			"internalGuard.nonceStore is nil; guard constructed without WithServiceTokenNonceStore"))
-	} else if kind := ns.Kind(); kind == auth.NonceStoreKindNoop {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneNonceStoreMissing,
-			"control-plane NonceStore must be a replay-safe implementation in "+
-				"adapter mode \"real\"; NoopNonceStore detected — inject "+
-				"InMemoryNonceStore (single pod) or a shared store (multi-pod) "+
-				"via WithServiceTokenNonceStore"))
-	} else if kind == auth.NonceStoreKindInMemory && !d.Topology.SinglePodReplayProtection && d.Topology.RequireProductionControlPlane() {
-		slog.Warn("controlplane: in-memory nonce store rejected for multi-pod deployment",
-			slog.String("nonce_store_kind", string(kind)),
-			slog.String("hint", "set GOCELL_SINGLE_POD=1 for single-pod deployments or configure a distributed NonceStore"))
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneNonceStoreMissing,
-			"in-memory nonce store requires GOCELL_SINGLE_POD=1 "+
-				"(single-pod deployments) or a distributed store via "+
-				"WithServiceTokenNonceStore (multi-pod); refuse fail-open"))
-	}
-	if requiresDistributedReplay(d.Topology) && d.ConsumerClaimerKind != consumerClaimerKindDistributed {
-		errs = append(errs, errcode.New(errcode.KindInternal, errcode.ErrControlplaneClaimerNotDistributed,
-			"ERR_CONTROLPLANE_CLAIMER_NOT_DISTRIBUTED: real multi-pod "+
-				"deployments require Redis-backed outbox idempotency claimer; "+
-				"set GOCELL_REDIS_ADDR or run with GOCELL_SINGLE_POD=1"))
-	}
-	return errs
-}
-
-func (d *SharedDeps) validateInternalListenerGuard() []error {
-	if d.InternalHTTPAddr == "" {
-		return []error{errcode.New(errcode.KindInvalid, errcode.ErrValidationFailed,
-			"SharedDeps.InternalHTTPAddr must be set; the internal listener is always enabled and protected by GOCELL_SERVICE_SECRET")}
-	}
-	if d.InternalGuard != nil {
-		return nil
-	}
-	return []error{errcode.New(errcode.KindInternal, errcode.ErrControlplaneServiceSecretMissing,
-		"SharedDeps.InternalGuard must be set to protect /internal/v1/*; set GOCELL_SERVICE_SECRET")}
-}
-
-func (d *SharedDeps) validateHealthReachability() []error {
-	if !d.Topology.RequireProductionControlPlane() || d.HealthLocalOnly {
-		return nil
-	}
-	if d.HealthHTTPAddr == "" || !isLoopbackBindAddr(d.HealthHTTPAddr) {
-		return nil
-	}
-	return []error{errcode.New(errcode.KindInvalid, errcode.ErrValidationFailed,
-		"GOCELL_HTTP_HEALTH_ADDR is loopback-only in adapter mode \"real\"; "+
-			"kubelet HTTP probes and Prometheus PodIP/Service scrapes cannot "+
-			"reach container loopback. Set GOCELL_HTTP_HEALTH_ADDR=:9091 "+
-			"(or a Pod-reachable address), or set GOCELL_HTTP_HEALTH_LOCAL_ONLY=1 "+
-			"only for same-pod sidecar or exec-probe deployments.")}
-}
-
-func isLoopbackBindAddr(addr string) bool {
-	host, _, err := net.SplitHostPort(addr)
-	if err != nil {
-		host = addr
-	}
-	host = strings.Trim(host, "[]")
-	if host == "" {
-		return false
-	}
-	if strings.EqualFold(host, "localhost") {
-		return true
-	}
-	ip := net.ParseIP(host)
-	return ip != nil && ip.IsLoopback()
-}
 
 // LoadSharedDepsFromEnv reads all environment variables and builds a fully
 // populated SharedDeps for cross-cutting concerns. Cell-specific dependencies
